dom/html/test/test_bug402680.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=402680
-->
<head>
  <title>Test for Bug 402680</title>
  <script>
    var activeElementIsNull = (document.activeElement == null);
  </script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=402680">Mozilla Bug 402680</a>
<p id="display"></p>
<div id="content">
  <input type="text">
  <textarea></textarea>
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 402680 **/

ok(activeElementIsNull,
   "Before document has body, active element should be null");

function testActiveElement() {
  ok(document.body == document.activeElement,
     "After page load body element should be the active element!");
  var input = document.getElementsByTagName("input")[0];
  input.focus();
  ok(document.activeElement == input,
     "Input element isn't the active element!");
  var textarea = document.getElementsByTagName("textarea")[0];
  textarea.focus();
  ok(document.activeElement == textarea,
     "Textarea element isn't the active element!");
  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();
addLoadEvent(testActiveElement);

</script>
</pre>
</body>
</html>
